Allow `download`, `download_and_extract`, and `file` in `module_ctx`
This involves some changes in logging and reporting, since a "download event" is not necessarily tied to a repository any more, but could be from a module extension impl function. So we generally convert any logging of repo names to a "context", which could either be a repo or a module extension.

Fixes https://github.com/bazelbuild/bazel/issues/16144

PiperOrigin-RevId: 469413473
Change-Id: I8c5b7477e6993a6fb77e07d82fc97260c19674b4